Finished overworld map tiles

+++ FinalQuestino/tile_map.c
@@ -23,6 +23,6 @@ void cTileMap_Init( cTileMap_t* map )
 
    for ( i = 0; i < PORTAL_COUNT; i++ )
    {
-      map->portals[i] = ( TILES_X * TILES_Y ) << 21; // off the map
+      map->portals[i] = (uint32_t)( TILES_X * TILES_Y ) << 21; // off the map
    }
 }
